add comments in serial and update execute time fn

+++ include/physics.h
@@ -92,42 +92,42 @@ heat_equation_2D_serial(double time_step, double time_limit,
 
 
 ////////////////heat equation 1D parallel execution time without i/o
-extern int
+extern double
 heat_equation_execution_time_1D_P1_MPI(double time_step, double time_limit,
                                        double space_step,
                                        int precision);
 
-extern int
+extern double
 heat_equation_execution_time_1D_P1_OPENMP(double time_step, double time_limit,
                                           double space_step,
                                           int precision);
 
-extern int
+extern double
 heat_equation_execution_time_1D_P1_OPENMP_V2(double time_step, double time_limit,
                                              double space_step,
                                              int precision);
 
 ////////////////heat equation 2D parallel execution time without i/o
-extern int
+extern double
 heat_equation_execution_time_2D_P1_MPI(double time_step, double time_limit,
                                        double spaceX_step, double spaceY_step,
                                        int precision);
 
-extern int heat_equation_execution_time_2D_P1_OPENMP(double time_step, double time_limit,
+extern double heat_equation_execution_time_2D_P1_OPENMP(double time_step, double time_limit,
                         double spaceX_step, double spaceY_step,
                         int precision);
 
-extern int heat_equation_execution_time_2D_P1_OPENMP_V2(double time_step, double time_limit,
+extern double heat_equation_execution_time_2D_P1_OPENMP_V2(double time_step, double time_limit,
                         double spaceX_step, double spaceY_step,
                         int precision);
 
 
 ////////////////heat equation 1D serial execution time without i/o
-extern int
+extern double
 heat_equation_execution_time_1D_serial(double time_step, double time_limit, double space_step, int precision);
 
 ////////////////heat equation 2D serial execution time without i/o
-extern int
+extern double
 heat_equation_execution_time_2D_serial(double time_step, double time_limit,
                                        double spaceX_step, double spaceY_step,
                                        int precision);
