+++ include/thermoserial.h
@@ -27,11 +27,11 @@
                         double space_step_y,
                         int precision);
 
-extern int
+extern double
 _execution_time_heat_transfer_1D_serial(double time_step, double time_limit,
                                         double space_step,
                                         int precision);
-extern int
+extern double
 _execution_time_heat_transfer_2D_serial(double time_step, double time_limit,
                                   double space_step_x,
                                   double space_step_y,
